spring5-mvc-rest-app/src/main/java/guru/springframework/controllers/v1/CategoryController.java
package guru.springframework.controllers.v1;


import guru.springframework.api.v1.model.CategoryDTO;
import guru.springframework.api.v1.model.CategoryListDTO;
import guru.springframework.services.CategoryService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

@RestController
@Slf4j
@RequestMapping("/api/v1/categories/")
public class CategoryController {

    private final CategoryService categoryService;
    public CategoryController(CategoryService categoryService) {
        this.categoryService = categoryService;
    }

    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    public CategoryListDTO getAllCategories() {

        CategoryListDTO list = categoryService.getAllCategories();
        log.info("list contains : {}", list.getCategoryDTOList().size());
        return list;

    }

    @GetMapping("{name}")
    @ResponseStatus(HttpStatus.OK)
    public CategoryDTO getCategoryByName(@PathVariable("name") String catName) {

        CategoryDTO cat = categoryService.getCategoryByName(catName);
        return cat;

    }
}
